Added reference anchors to pages
Reference anchors are now added to the top of our .rst pages and are
set to the convention of using the page header as the anchor text
(lower case and hyphenated). For example, an .rst file with  a page
header of "Reference Me" would have a reference string as follows:
":ref:`reference-me`.